Pull CLI helper functions out to a library.
FIXME: The tsc config changes here are only temporary until cli.js is
ported to typescript. In particular, `"declaration": true` should be
restored after that's done.
