+++ tsconfig.json
@@ -3,8 +3,9 @@
     "target": "es6",
     "module": "commonjs",
     "moduleResolution": "node",
+    "allowJs": true,
+    "checkJs": true,
     "sourceMap": true,
-    "declaration": true,
     "noImplicitAny": false,
     "rootDir": "./src",
     "outDir": "./dist",
@@ -18,6 +19,7 @@
   },
   "exclude": [
     "node_modules",
-    "dist"
+    "dist",
+    "bin"
   ]
 }
